requirements.txt
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-google-genai
chromadb
spacy
python-dotenv
llama-index-llms-google-genai